requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.12.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==41.0.8
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.104.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
groq==0.4.1
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.25.2
idna==3.10
isodate==0.7.2
lxml==5.4.0
more-itertools==10.7.0
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rsa==4.9.1
simple-salesforce==1.12.4
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.8.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn[standard]==0.24.0
zeep==4.3.1

# Testing and development
pytest==7.4.3
pytest-asyncio==0.21.1
coverage==7.3.2

# Code quality and formatting
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Performance and monitoring
aiofiles==23.2.1

# Documentation (optional)
mkdocs==1.5.3
mkdocs-material==9.4.8
